requirements.txt
opencv-python==4.8.1.78
numpy==1.26.0
pillow==10.0.0
pygame==2.5.2
moviepy==1.0.3 